requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# webSMSSender-master/app.py: 1,2
Flask == 1.1.2

# webSMSSender-master/app.py: 3
Pyrebase4 == 4.5.0

# webSMSSender-master/app.py: 5
requests == 2.25.1
gunicorn